refactor Next_Steps
